version where issue 10 (update info) is resolved

+++ trunk/imagefs.py
@@ -29,9 +29,12 @@ class ImageFS(Fuse):
         
     def getattr(self, path):
         self.dirmaker.makeDirs(self.src)
-        if os.path.isdir(self.pathTempDir+self.src+ path) and not os.path.isdir(self.src+path):
-            shutil.rmtree(self.pathTempDir+self.src+ path)
-            
+        
+        if path.split("/")[len(path.split("/"))-1] not in self.dirmaker.getDirNames():
+            if os.path.isdir(self.pathTempDir+self.src+ path) and not os.path.isdir(self.src+path):
+                shutil.rmtree(self.pathTempDir+self.src+path)
+                    
+        
         for dirName in self.dirmaker.getDirNames():
             if len(path.split("/"+dirName)) > 1:
                 if os.path.isfile(self.pathTempDir+self.src+ path):
